Moved random bot to folder
